requirements.txt
groq
python-telegram-bot
requests
feedparser
python-dotenv
huggingface_hub